requirements.txt
discord.py==1.6.0
pylint==2.6.0
pytest==6.2.2
python-dotenv==0.15.0